+++ tests/test_templates.py
@@ -1,6 +1,4 @@
 import os
-import pathlib
-from contextlib import nullcontext
 
 import pytest
 
@@ -11,7 +9,6 @@ from esmerald.requests import Request
 from esmerald.routing.gateways import Gateway
 from esmerald.routing.handlers import get
 from esmerald.template.jinja import JinjaTemplateEngine
-from esmerald.template.mako import MakoTemplateEngine
 from esmerald.testclient import EsmeraldTestClient, create_client
 
 
@@ -27,20 +24,6 @@ def test_handler_raise_for_no_template_engine_created() -> None:
             "detail": "Template engine is not configured",
         }
         assert response.status_code == 500
-
-
-@pytest.mark.parametrize("engine", [JinjaTemplateEngine, MakoTemplateEngine])
-def test_engine_jinja_and_mako(engine, template_dir: "pathlib.Path") -> None:
-    with pytest.warns(DeprecationWarning) if engine is MakoTemplateEngine else nullcontext():
-        app = Esmerald(
-            routes=[],
-            template_config=TemplateConfig(
-                directory=template_dir,
-                engine=engine,
-            ),
-        )
-
-    assert isinstance(app.template_engine, engine)
 
 
 @pytest.mark.parametrize("apostrophe", ["'", '"'])